requirements.txt
# Core Dependencies
fastapi>=0.100.0
uvicorn>=0.20.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# LangChain & AI
langchain>=0.0.300
langchain-core>=0.1.0
langchain-community>=0.2.17
langchain-openai>=0.1.25
langchain-chroma>=0.0.1
langchain-milvus>=0.1.5
langchain-ollama>=0.1.3
openai>=1.52.0
tiktoken>=0.7.0

# Vector Database & Search
chromadb>=0.4.0
sqlalchemy>=2.0.0
rank-bm25>=0.2.2

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
python-multipart>=0.0.6
beautifulsoup4>=4.12.3
openpyxl>=3.1.0

# UI/Frontend
streamlit>=1.24.0
streamlit-chat==0.1.1


# Utils
requests>=2.31.0
tqdm>=4.65.0
PyYAML>=6.0.0

# New additions
pymilvus>=2.3.1
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.5.0


transformers>=4.30.0
sentence-transformers

